Refactor #36 - Move HomeStateMachineFactory to test
