+++ app/src/test/java/com/nlab/reminder/domain/feature/home/HomeStateMachineFactoryTest.kt
@@ -1,0 +1,51 @@
+/*
+ * Copyright (C) 2022 The N's lab Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.nlab.reminder.domain.feature.home
+
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import org.hamcrest.CoreMatchers.*
+import org.hamcrest.MatcherAssert.*
+import org.junit.Test
+import org.mockito.kotlin.mock
+
+
+/**
+ * @author Doohyun
+ */
+class HomeStateMachineFactoryTest {
+    @Test
+    fun `holds init state when machine created by factory`() {
+        val stateMachineFactory = HomeStateMachineFactory(
+            getHomeSummary = mock(),
+            getTagUsageCount = mock(),
+            modifyTagName = mock(),
+            deleteTag = mock()
+        )
+
+        assertThat(
+            stateMachineFactory
+                .create(
+                    scope = CoroutineScope(Dispatchers.Default),
+                    navigationEffect = mock()
+                )
+                .state
+                .value,
+            equalTo(HomeState.Init)
+        )
+    }
+}
